+++ .changeset/tame-coats-wink.md
@@ -1,0 +1,5 @@
+---
+"ingred-ui": major
+---
+
+Replace NewDatePicker with DatePicker and drop the old DatePicker. Replace NewDateRangePicker with DateRangePicker and drop the old DateRangePicker. Drop moment.js and react-dates.
