+++ package.json
@@ -28,8 +28,6 @@
   "dependencies": {
     "@floating-ui/react": "^0.25.1",
     "dayjs": "^1.11.6",
-    "moment": "^2.29.3",
-    "react-dates": "^21.8.0",
     "react-select": "^5.7.0",
     "react-transition-group": "^4.4.1"
   },
@@ -58,7 +56,6 @@
     "@testing-library/react": "14.0.0",
     "@types/jest": "29.5.1",
     "@types/react": "18.0.12",
-    "@types/react-dates": "21.8.3",
     "@types/react-dom": "18.0.5",
     "@types/react-transition-group": "4.4.4",
     "@types/rollup-plugin-peer-deps-external": "2.2.1",
