+++ src/components/DatePicker/DatePicker.tsx
@@ -1,114 +1,155 @@
-import * as React from "react";
-import * as Styled from "./styled";
-import "react-dates/initialize";
-import dayjs, { InstanceLocaleDataReturn } from "dayjs";
-import moment from "moment";
+import React, { forwardRef, memo, useCallback, useState } from "react";
+import { Flex, Calendar, DateField } from "..";
+import { Action } from "../Calendar/internal/Actions";
+import { Dayjs } from "dayjs";
 import {
-  dayjsToMoment,
-  momentToDayjs,
-  convertDayjsLocaleDataToObject,
-} from "../../utils/time";
-import {
-  RenderMonthProps,
-  SingleDatePicker,
-  SingleDatePickerShape,
-} from "react-dates";
-import Icon from "../Icon";
-import Spacer from "../Spacer";
+  flip,
+  offset,
+  useDismiss,
+  useFloating,
+  useInteractions,
+  useRole,
+} from "@floating-ui/react";
 import { useLocaleProps } from "../../hooks/useLocaleProps";
 
-function isOutsideRange() {
-  return false;
-}
+export type DatePickerProps = {
+  /**
+   * 日付
+   * @default dayjs()
+   */
+  date: Dayjs;
+  /**
+   * デフォルトで選択されているアクション
+   */
+  defaultClickAction?: string;
+  /**
+   * カレンダーの左に表示するアクション
+   */
+  actions?: Action[];
+  /**
+   * アクションをクリックしたときの挙動
+   */
+  onClickAction?: (action: Action) => void;
+  /**
+   * 指定したい format
+   * @default YYYY-MM-DD
+   */
+  format?: string;
+  /**
+   * エラーメッセージのテキスト
+   */
+  errorText?: string;
+  /**
+   * 入力を無効にする
+   * @default false
+   */
+  disabled?: boolean;
+  /**
+   * カレンダーに表示する年月のフォーマット
+   */
+  monthFormat?: string;
+  /**
+   * 選択可能なカレンダーの領域を制限する
+   * true が返る場合は、選択不可となる
+   * @default () => false
+   */
+  isOutsideRange?: (date: Dayjs) => boolean;
+  /**
+   * 日付が変更されたときに呼ばれる関数
+   */
+  onDateChange: (date: Dayjs) => void;
+};
 
-export type DatePickerProps = Partial<
-  Omit<
-    SingleDatePickerShape,
-    "date" | "onFocusChange" | "onDateChange" | "renderMonthText"
-  >
-> &
-  // MEMO: Add RenderMonthProps to pass type check.
-  Omit<RenderMonthProps, "renderMonthText"> & {
-    date: dayjs.Dayjs | null;
-    onDateChange: (date: dayjs.Dayjs | null) => void;
-    renderMonthText?: ((month: dayjs.Dayjs) => React.ReactNode) | null;
-    locale?: string;
-    localeData?: InstanceLocaleDataReturn;
-    error?: boolean;
-  };
-
-const DatePicker = React.forwardRef<HTMLDivElement, DatePickerProps>(
+/**
+ * @memo 次のメジャーリリースで DatePicker に変更。現行の DatePicker は削除。
+ */
+export const NewDatePicker = forwardRef<HTMLDivElement, DatePickerProps>(
   function DatePicker(inProps, ref) {
     const props = useLocaleProps({ props: inProps, name: "DatePicker" });
     const {
       date,
-      error = false,
+      format = "YYYY-MM-DD",
+      disabled = false,
+      monthFormat,
+      weekList,
+      isOutsideRange,
+      errorText,
+      defaultClickAction,
+      actions,
+      onClickAction,
       onDateChange,
-      renderMonthText,
-      renderMonthElement,
-      locale = "en",
-      localeData,
-      ...rest
     } = props;
 
-    const [focused, setFocused] = React.useState<boolean>(false);
-    const onFocusChange = ({ focused }: { focused: boolean }) => {
-      setFocused(focused);
-    };
-    const handleDateChange = (date: moment.Moment | null) => {
-      const dayjsize = momentToDayjs(date);
-      onDateChange(dayjsize);
-    };
-    const handleRenderMonthText = (month: moment.Moment) => {
-      if (renderMonthText == undefined) return;
-      const dayjsize = momentToDayjs(month);
-      if (dayjsize === null) return;
-      return renderMonthText(dayjsize);
-    };
+    const [open, setOpen] = useState(false);
+    const { context, refs, strategy, x, y } = useFloating({
+      placement: "right-start",
+      open,
+      onOpenChange: setOpen,
+      middleware: [offset(10), flip()],
+    });
 
-    if (localeData) {
-      moment.locale(locale, convertDayjsLocaleDataToObject(localeData));
-    } else {
-      moment.locale(locale);
-    }
+    const dismiss = useDismiss(context);
+    const role = useRole(context);
+
+    const { getReferenceProps, getFloatingProps } = useInteractions([
+      dismiss,
+      role,
+    ]);
+
+    const handleClickCalendarIcon = useCallback(() => {
+      if (disabled) return;
+      setOpen((prev) => !prev);
+    }, [disabled]);
+
+    const handleClickDate = useCallback(
+      (date: Dayjs) => {
+        onDateChange(date);
+        setOpen(false);
+      },
+      [onDateChange],
+    );
+
+    const handleClose = useCallback(() => {
+      setOpen(false);
+    }, []);
 
     return (
-      <Styled.Container ref={ref} error={error}>
-        <SingleDatePicker
-          id="datePicker"
-          focused={focused}
-          date={dayjsToMoment(date)}
-          isOutsideRange={isOutsideRange}
-          numberOfMonths={1}
-          enableOutsideDays={true}
-          daySize={41}
-          weekDayFormat="ddd"
-          hideKeyboardShortcutsPanel={true}
-          navPrev={
-            <Styled.NavPrev>
-              <Spacer p={0.5}>
-                <Icon name="arrow_left" size="md" />
-              </Spacer>
-            </Styled.NavPrev>
-          }
-          navNext={
-            <Styled.NavNext>
-              <Spacer p={0.5}>
-                <Icon name="arrow_right" size="md" />
-              </Spacer>
-            </Styled.NavNext>
-          }
-          renderMonthText={
-            renderMonthText ? handleRenderMonthText : renderMonthText
-          }
-          renderMonthElement={renderMonthElement as never}
-          onFocusChange={onFocusChange}
-          onDateChange={handleDateChange}
-          {...rest}
-        />
-      </Styled.Container>
+      <Flex ref={ref} style={{ width: "fit-content" }}>
+        <div ref={refs.setReference} {...getReferenceProps()}>
+          <DateField
+            date={date}
+            format={format}
+            errorText={errorText}
+            disabled={disabled}
+            onClick={handleClickCalendarIcon}
+            onDateChange={onDateChange}
+          />
+        </div>
+        {open && (
+          <Calendar
+            ref={refs.setFloating}
+            date={date}
+            monthFormat={monthFormat}
+            weekList={weekList}
+            defaultClickAction={defaultClickAction}
+            actions={actions}
+            isOutsideRange={isOutsideRange}
+            style={{
+              position: strategy,
+              top: y ?? 0,
+              left: x ?? 0,
+              zIndex: 100,
+              overflow: "hidden",
+            }}
+            onClickAction={onClickAction}
+            onClickCloseButton={handleClose}
+            onDateChange={handleClickDate}
+            {...getFloatingProps()}
+          />
+        )}
+      </Flex>
     );
   },
 );
 
-export default DatePicker;
+export default memo(NewDatePicker);
