--- src/components/DatePicker/__tests__/DatePicker.test.tsx
@@ -1,16 +1,0 @@
-import * as React from "react";
-import "@testing-library/jest-dom";
-import { cleanup } from "@testing-library/react";
-import DatePicker from "..";
-import { renderWithThemeProvider } from "../../../utils/renderWithThemeProvider";
-
-describe("DatePicker component testing", () => {
-  afterEach(cleanup);
-
-  test("DatePicker", () => {
-    const { asFragment } = renderWithThemeProvider(
-      <DatePicker date={null} onDateChange={jest.fn()} />,
-    );
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
