+++ src/components/DateRangePicker/DateRangePicker.tsx
@@ -1,126 +1,177 @@
 import * as React from "react";
-import * as Styled from "./styled";
-import "react-dates/initialize";
-import dayjs, { InstanceLocaleDataReturn } from "dayjs";
-import moment from "moment";
+import { Flex, DateRangeField, CalendarRange } from "..";
+import { forwardRef, useRef, useState } from "react";
 import {
-  dayjsToMoment,
-  momentToDayjs,
-  convertDayjsLocaleDataToObject,
-} from "../../utils/time";
+  flip,
+  offset,
+  useDismiss,
+  useFloating,
+  useInteractions,
+  useRole,
+} from "@floating-ui/react";
+import { Dayjs } from "dayjs";
+import { Action } from "../Calendar/internal/Actions";
 import {
-  FocusedInputShape,
-  DateRangePicker as OriginalDateRangePicker,
-  DateRangePickerShape,
-} from "react-dates";
-import Icon from "../Icon";
-import Spacer from "../Spacer";
+  ClickState,
+  ClickStateType,
+} from "../Calendar/CalendarRange/constants";
+import { useMergeRefs } from "../../hooks/useMergeRefs";
 import { useLocaleProps } from "../../hooks/useLocaleProps";
 
-function isOutsideRange() {
-  return false;
-}
-
-export type DateRangePickerProps = Partial<
-  Omit<
-    DateRangePickerShape,
-    "startDate" | "endDate" | "onDatesChange" | "renderMonthText"
-  >
-> & {
-  startDate: dayjs.Dayjs | null;
-  endDate: dayjs.Dayjs | null;
-  onDatesChange: (arg: {
-    startDate: dayjs.Dayjs | null;
-    endDate: dayjs.Dayjs | null;
-  }) => void;
-  renderMonthText?: ((month: dayjs.Dayjs) => React.ReactNode) | null;
-  locale?: string;
-  localeData?: InstanceLocaleDataReturn;
-  error?: boolean;
+export type DateRange = {
+  startDate: Dayjs;
+  endDate: Dayjs;
 };
 
-const DateRangePicker = React.forwardRef<HTMLDivElement, DateRangePickerProps>(
-  function DateRangePicker(inProps, ref) {
-    const props = useLocaleProps({ props: inProps, name: "DateRangePicker" });
+export type DateRangePickerProps = {
+  /**
+   * 開始日
+   * @default dayjs()
+   */
+  startDate: Dayjs;
+  /**
+   * 終了日
+   * @default dayjs()
+   */
+  endDate: Dayjs;
+  /**
+   * エラーメッセージのテキスト
+   */
+  errorText?: string;
+  /**
+   * デフォルトで選択されているアクション
+   */
+  defaultClickAction?: string;
+  /**
+   * カレンダーの左に表示するアクション
+   */
+  actions?: Action[];
+  /**
+   * アクションをクリックしたときの挙動
+   */
+  onClickAction?: (action: Action) => void;
+  /**
+   * 入力を無効にする
+   * @default false
+   */
+  disabled?: boolean;
+  /**
+   * カレンダーに表示する年月のフォーマット
+   */
+  monthFormat?: string;
+  /**
+   * 選択可能なカレンダーの領域を制限する
+   * true が返る場合は、選択不可となる
+   * @default () => false
+   */
+  isOutsideRange?: (date: Dayjs) => boolean;
+  /**
+   * 日付が変更されたときに呼ばれる関数
+   */
+  onDatesChange: (dates: DateRange) => void;
+};
+
+/**
+ * @memo 次のメジャーリリースで DateRangePicker に変更。現行の DateRangePicker は削除。
+ */
+export const DateRangePicker = forwardRef<HTMLDivElement, DateRangePickerProps>(
+  function DateRangePicker(inProps, propRef) {
+    const props = useLocaleProps({
+      props: inProps,
+      name: "DateRangePicker",
+    });
     const {
       startDate,
       endDate,
-      error = false,
+      errorText,
+      disabled = false,
+      monthFormat,
+      weekList,
+      isOutsideRange = () => false,
+      defaultClickAction,
+      actions,
+      onClickAction,
       onDatesChange,
-      renderMonthText,
-      renderMonthElement,
-      locale = "en",
-      localeData,
-      ...rest
     } = props;
-    const [focusedInput, setFocusedInput] =
-      React.useState<FocusedInputShape | null>(null);
 
-    const handleDatesChange = (arg: {
-      startDate: moment.Moment | null;
-      endDate: moment.Moment | null;
-    }) => {
-      const dayjsize = {
-        startDate: momentToDayjs(arg.startDate),
-        endDate: momentToDayjs(arg.endDate),
-      };
-      onDatesChange(dayjsize);
-    };
-    const handleRenderMonthText = (month: moment.Moment) => {
-      if (renderMonthText == undefined) return;
-      const dayjsize = momentToDayjs(month);
-      if (dayjsize === null) return;
-      return renderMonthText(dayjsize);
+    const [open, setOpen] = useState(false);
+    const { context, refs, strategy, x, y } = useFloating({
+      placement: "right-start",
+      open,
+      onOpenChange: setOpen,
+      middleware: [offset(10), flip()],
+    });
+    const innerRef = useRef<HTMLDivElement>(null);
+    const ref = useMergeRefs(propRef, innerRef);
+
+    const dismiss = useDismiss(context);
+    const role = useRole(context);
+
+    const { getReferenceProps, getFloatingProps } = useInteractions([
+      dismiss,
+      role,
+    ]);
+
+    const handleClickCalendarIcon = () => {
+      if (disabled) return;
+      setOpen((prev) => !prev);
     };
 
-    if (localeData) {
-      moment.locale(locale, convertDayjsLocaleDataToObject(localeData));
-    } else {
-      moment.locale(locale);
-    }
+    const handleClose = (clickState: ClickStateType) => {
+      if (clickState === ClickState.END) {
+        setOpen(false);
+      }
+    };
+
+    const handleClickCloseButton = React.useCallback(() => {
+      setOpen(false);
+    }, []);
 
     return (
-      <Styled.Container ref={ref} error={error}>
-        <OriginalDateRangePicker
-          startDatePlaceholderText="FROM"
-          endDatePlaceholderText="TO"
-          isOutsideRange={isOutsideRange}
-          customArrowIcon={<Styled.CustomArrowIcon />}
-          numberOfMonths={2}
-          enableOutsideDays={true}
-          daySize={41}
-          weekDayFormat="ddd"
-          hideKeyboardShortcutsPanel={true}
-          startDateId="startDate"
-          endDateId="endDate"
-          navPrev={
-            <Styled.NavPrev>
-              <Spacer p={0.5}>
-                <Icon name="arrow_left" size="md" />
-              </Spacer>
-            </Styled.NavPrev>
-          }
-          navNext={
-            <Styled.NavNext>
-              <Spacer p={0.5}>
-                <Icon name="arrow_right" size="md" />
-              </Spacer>
-            </Styled.NavNext>
-          }
-          startDate={dayjsToMoment(startDate)}
-          endDate={dayjsToMoment(endDate)}
-          renderMonthText={
-            renderMonthText ? handleRenderMonthText : renderMonthText
-          }
-          renderMonthElement={renderMonthElement as never}
-          focusedInput={focusedInput}
-          onFocusChange={setFocusedInput}
-          onDatesChange={handleDatesChange}
-          {...rest}
-        />
-      </Styled.Container>
+      <Flex ref={ref}>
+        <div
+          ref={refs.setReference}
+          {...getReferenceProps()}
+          style={{ width: "fit-content" }}
+        >
+          <DateRangeField
+            date={{
+              startDate,
+              endDate,
+            }}
+            errorText={errorText}
+            disabled={disabled}
+            onDatesChange={onDatesChange}
+            onClickCalendarIcon={handleClickCalendarIcon}
+          />
+        </div>
+        {open && (
+          <CalendarRange
+            ref={refs.setFloating}
+            startDate={startDate}
+            endDate={endDate}
+            defaultClickAction={defaultClickAction}
+            actions={actions}
+            monthFormat={monthFormat}
+            weekList={weekList}
+            style={{
+              position: strategy,
+              top: y ?? 0,
+              left: x ?? 0,
+              zIndex: 100,
+              overflow: "hidden",
+            }}
+            isOutsideRange={isOutsideRange}
+            onClickAction={onClickAction}
+            onClose={handleClose}
+            onClickCloseButton={handleClickCloseButton}
+            onDatesChange={onDatesChange}
+            {...getFloatingProps()}
+          />
+        )}
+      </Flex>
     );
   },
 );
 
-export default DateRangePicker;
+export default React.memo(DateRangePicker);
