+++ src/components/DateRangePicker/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./DateRangePicker";
+export type { DateRangePickerProps, DateRange } from "./DateRangePicker";
