--- src/components/DateRangePicker/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./DateRangePicker";
-export type { DateRangePickerProps } from "./DateRangePicker";
