--- src/components/NewDatePicker/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./NewDatePicker";
-export type { NewDatePickerProps } from "./NewDatePicker";
