--- src/components/NewDateRangePicker/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./NewDateRangePicker";
-export type { NewDateRangePickerProps, DateRange } from "./NewDateRangePicker";
